refactor: Move util functions
